Refatorando Toasts
