+++ src/components/ToastContainer/index.tsx
@@ -1,31 +1,18 @@
 import React from 'react';
-import { FiAlertCircle, FiXCircle } from 'react-icons/fi';
 
-import { Container, Toast } from './styles'
+import { Container } from './styles'
 
-import { ToastMessage, useToast } from '../../hooks/toasts'
-
+import { ToastMessage, useToast} from '../../hooks/toasts'
+import Toast from './Toast'
 interface ToastContainerProps {
     messages: ToastMessage[];
 }
 
 const ToastContainer: React.FC<ToastContainerProps> = ({ messages }) => {
-    const { removeToast } = useToast();
-
     return (
         <Container>
           {messages.map((message) => (
-            <Toast key={message.id} type={message.type}>
-              <FiAlertCircle size={20} />
-                <div>
-                    <strong> {message.title} </strong>
-                    <p> {message.description} </p>
-                </div>
-
-                <button onClick={() => removeToast(message.id)} type="button">
-                    <FiXCircle size={18} />
-                </button>
-            </Toast>
+            <Toast key={message.id} message={message} />
           ))}
         </Container>
     );
